refactor(sharing-dialog): rename context to fetching-context
