+++ components/sharing-dialog/src/access-list/list-item.js
@@ -4,8 +4,8 @@ import { useOnlineStatus } from '@dhis2/app-runtime'
 import { colors } from '@dhis2/ui-constants'
 import PropTypes from 'prop-types'
 import React, { useContext } from 'react'
-import { FetchingContext } from '../context/index.js'
 import { DestructiveSelectOption } from '../destructive-select-option/index.js'
+import { FetchingContext } from '../fetching-context/index.js'
 import { isRemovableTarget } from '../helpers/index.js'
 import i18n from '../locales/index.js'
 import { ListItemContext } from './list-item-context.js'
